requirement.txt
flask 
gunicorn
textblob
google-generativeai